feat(integration/notion): introduce comment created definition and moved base payload schema
